Updating handler
